LIME-1545 - Removing overload protection and adding protectConfig override
