+++ .secrets.baseline
@@ -285,5 +285,5 @@
       }
     ]
   },
-  "generated_at": "2025-02-03T09:05:23Z"
+  "generated_at": "2025-03-14T17:43:21Z"
 }
